requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
openai==1.12.0
faiss-cpu==1.7.4
numpy==1.24.3
python-jose==3.3.0
cryptography==41.0.7
SQLAlchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
prometheus-client==0.19.0
aiohttp==3.9.3
pydantic==2.5.3
python-i18n==0.3.9
sentence-transformers==2.3.1
asyncpg==0.28.0 